+++ shake-ats/CHANGELOG.md
@@ -1,5 +1,9 @@
 # shake-ats
 
+## 1.10.0.0
+
+  * Make `cabalForeign` take a `HsCompiler` instead of a `CCompiler`.
+
 ## 1.9.0.5
 
   * Bugfix for recent version of `shake-cabal`