+++ shake-cabal/CHANGELOG.md
@@ -1,5 +1,9 @@
 # shake-cabal
 
+## 0.2.0.0
+
+  * More precise `HsCompiler` type
+
 ## 0.1.0.5
 
   * Slightly better behavior for `getCabalDepsA` and `getCabalDepsV`